moved models into the secret_holder model file
